use file-scoped namespaces in data handlers, move SiteLocaleDataSourceHandler to DataSourceHandlers/Locale
