+++ Apps.Webflow/DataSourceHandlers/Locale/UpdateCollectionItemLocaleDataSourceHandler.cs
@@ -2,13 +2,12 @@
 using Blackbird.Applications.Sdk.Common;
 using Blackbird.Applications.Sdk.Common.Invocation;
 
-namespace Apps.Webflow.DataSourceHandlers.Locale
+namespace Apps.Webflow.DataSourceHandlers.Locale;
+
+public class UpdateCollectionItemLocaleDataSourceHandler : CmsLocaleDataSourceHandler
 {
-    public class UpdateCollectionItemLocaleDataSourceHandler : CmsLocaleDataSourceHandler
+    public UpdateCollectionItemLocaleDataSourceHandler(InvocationContext invocationContext,
+        [ActionParameter] UpdateCollectionItemRequest request) : base(invocationContext, request.SiteId)
     {
-        public UpdateCollectionItemLocaleDataSourceHandler(InvocationContext invocationContext,
-            [ActionParameter] UpdateCollectionItemRequest request) : base(invocationContext, request.SiteId)
-        {
-        }
     }
 }
