+++ Apps.Webflow/DataSourceHandlers/PageDataSourceHandler.cs
@@ -6,35 +6,34 @@ using Blackbird.Applications.Sdk.Common.Dynamic;
 using Blackbird.Applications.Sdk.Common.Invocation;
 using RestSharp;
 
-namespace Apps.Webflow.DataSourceHandlers
+namespace Apps.Webflow.DataSourceHandlers;
+
+public class PageDataSourceHandler : WebflowInvocable, IAsyncDataSourceItemHandler
 {
-    public class PageDataSourceHandler : WebflowInvocable, IAsyncDataSourceItemHandler
+    private readonly GetPageAsHtmlRequest _input;
+
+    public PageDataSourceHandler(InvocationContext invocationContext, [ActionParameter] GetPageAsHtmlRequest input):base(invocationContext)
     {
-        private readonly GetPageAsHtmlRequest _input;
+        _input = input;
+    }
 
-        public PageDataSourceHandler(InvocationContext invocationContext, [ActionParameter] GetPageAsHtmlRequest input):base(invocationContext)
-        {
-            _input = input;
-        }
+    public async Task<IEnumerable<DataSourceItem>> GetDataAsync(DataSourceContext context, CancellationToken cancellationToken)
+    {
+        var siteId = _input.SiteId;
 
-        public async Task<IEnumerable<DataSourceItem>> GetDataAsync(DataSourceContext context, CancellationToken cancellationToken)
-        {
-            var siteId = _input.SiteId;
+        var endpoint = $"sites/{siteId}/pages";
+        var request = new RestRequest(endpoint, Method.Get);
 
-            var endpoint = $"sites/{siteId}/pages";
-            var request = new RestRequest(endpoint, Method.Get);
+        var response = await Client.ExecuteWithErrorHandling<ListPagesResponse>(request);
 
-            var response = await Client.ExecuteWithErrorHandling<ListPagesResponse>(request);
+        var dataSourceItems = response.Pages
+            .Select(page => new DataSourceItem
+            {
+                Value = page.Id,
+                DisplayName = $"{page.Title}"
+            })
+            .ToList();
 
-            var dataSourceItems = response.Pages
-                .Select(page => new DataSourceItem
-                {
-                    Value = page.Id,
-                    DisplayName = $"{page.Title}"
-                })
-                .ToList();
-
-            return dataSourceItems;
-        }
+        return dataSourceItems;
     }
 }
