+++ Tests.Webflow/DataSources.cs
@@ -1,6 +1,7 @@
 ﻿using Apps.Webflow.DataSourceHandlers;
 using Apps.Webflow.DataSourceHandlers.Collection;
 using Apps.Webflow.DataSourceHandlers.CollectionItem;
+using Apps.Webflow.DataSourceHandlers.Locale;
 using Apps.Webflow.Models.Request.Components;
 using Apps.Webflow.Models.Request.Pages;
 using Blackbird.Applications.Sdk.Common.Dynamic;
